refactor(istd): refrectored to multiple files
