+++ iscouttaskdispatch/site/templates/base.html
@@ -10,8 +10,8 @@
 <body>
     <div class="navbar">
         <a href="{{ url_for('site.index') }}">Home</a>
-        <a href="{{ url_for('site.showAllTasks') }}">Tasks</a>
-        <a href="{{ url_for('site.showAllTeams') }}">Teams</a>
+        <a href="{{ url_for('site.tasks_site.showAllTasks') }}">Tasks</a>
+        <a href="{{ url_for('site.teams_site.showAllTeams') }}">Teams</a>
         <!-- Add more navigation links as needed -->
     </div>
     <div class="content">
