+++ iscouttaskdispatch/site/templates/index.html
@@ -7,6 +7,6 @@
     <h1>iScout Aufgabenverteilung</h1>
     <p>Main Page</p>
     <p>All Tasks:</p>
-    <a href="{{ url_for('site.showAllTasks') }}" class="button">View Tasks</a>
+    <a href="{{ url_for('site.tasks_site.showAllTasks') }}" class="button">View Tasks</a>
 </div>
 {% endblock %}
