+++ iscouttaskdispatch/site/templates/task/delete.html
@@ -8,7 +8,7 @@
 <div class="container">
     <h1>Delete Task</h1>
     <p>Enter the ID of the task you want to delete:</p>
-    <form action="{{ url_for('site.deleteTaskSite') }}" method="post">
+    <form action="{{ url_for('site.tasks_site.deleteTaskSite') }}" method="post">
         <label for="taskID">Task ID:</label>
         <input type="number" id="taskID" name="taskID" required>
         <br>
@@ -20,6 +20,6 @@
         </div>
     {% endif %}
     <br>
-    <a href="{{ url_for('site.showAllTasks') }}">Back to Task List</a>
+    <a href="{{ url_for('site.tasks_site.showAllTasks') }}">Back to Task List</a>
 </div>
 {% endblock %}
