+++ iscouttaskdispatch/site/templates/team/create.html
@@ -5,7 +5,7 @@
 {% block content %}
     <div class="container">
         <h1>Create team</h1>
-        <form action="{{ url_for('site.createTeamSite') }}" method="post">
+        <form action="{{ url_for('site.teams_site.createTeamSite') }}" method="post">
             <label for="teamID">team ID:</label>
             <input type="number" id="teamID" name="teamID" value="{{ form_data['teamID'] }}" required>
             <br>
@@ -21,6 +21,6 @@
             </div>
         {% endif %}
         <br>
-        <a href="{{ url_for('site.showAllTeams') }}">Back to team List</a>
+        <a href="{{ url_for('site.teams_site.showAllTeams') }}">Back to team List</a>
     </div>
 {% endblock %}
