+++ iscouttaskdispatch/site/templates/team/delete.html
@@ -8,7 +8,7 @@
 <div class="container">
     <h1>Delete Team</h1>
     <p>Enter the ID of the team you want to delete:</p>
-    <form action="{{ url_for('site.deleteTeamSite') }}" method="post">
+    <form action="{{ url_for('site.teams_site.deleteTeamSite') }}" method="post">
         <label for="teamID">Team ID:</label>
         <input type="number" id="teamID" name="teamID" required>
         <br>
@@ -20,6 +20,6 @@
         </div>
     {% endif %}
     <br>
-    <a href="{{ url_for('site.showAllTeams') }}">Back to Team List</a>
+    <a href="{{ url_for('site.teams_site.showAllTeams') }}">Back to Team List</a>
 </div>
 {% endblock %}
