+++ iscouttaskdispatch/site/templates/team/edit.html
@@ -4,7 +4,7 @@
 
 {% block content %}
     <h1>Edit Team {{ team.teamID }}</h1>
-    <form action="{{ url_for('site.updateTeam', teamID=team.teamID) }}" method="post">
+    <form action="{{ url_for('site.teams_site.updateTeam', teamID=team.teamID) }}" method="post">
         <label for="name">Team Name:</label>
         <input type="text" id="name" name="name" value="{{ team.name }}">
         <br>
