+++ iscouttaskdispatch/site/templates/team/index.html
@@ -5,8 +5,8 @@
 {% block content %}
 <div>
     <div class="navbar">
-        <a href="{{ url_for('site.createTeamSite') }}">Create New</a>
-        <a href="{{ url_for('site.deleteTeamSite') }}">Delete</a>
+        <a href="{{ url_for('site.teams_site.createTeamSite') }}">Create New</a>
+        <a href="{{ url_for('site.teams_site.deleteTeamSite') }}">Delete</a>
     </div>
     <div class="page-div">
         <h1>Teams</h1>
